# agrotech_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AgroTech",
        "version": "1.0.0",
        "title": "AgroTech",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yogeshwaran@doodlblue.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:2122",
    "basePath": "/api/v1",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/user/deleteAdmin": {
            "delete": {
                "tags": [
                    "DeleteAdmin"
                ],
                "summary": "response",
                "description": "DeleteRole",
                "operationId": "DeleteRole",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "response",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                "message": "User Deleted Successfully",
                                "data": null
                            }
                        }
                    }
                }
            }
        },
...
        "/roles/addRole": {
            "post": {
                "tags": [
                    "AddRole"
                ],
                "summary": "response",
                "description": "AddRole",
                "operationId": "AddRole",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "response",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/addRole"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                    "message": "Role Created Successfully",
                                    "data": {
                                        "permissions": [],
                                        "_id": "6235a79a78b7bf9e82cbcecd",
                                        "roleName": "Customer",
                                        "description": "Customer",
                                        "roleId": "ROLE_6",
                                        "createdAt": "2022-03-19T09:51:22.906Z",
                                        "updatedAt": "2022-03-19T09:51:22.906Z",
                                        "__v": 0
                                    }
                            }
                        }
                    }
                }
            }
        },
...
        "/roles/deleteRole": {
            "delete": {
                "tags": [
                    "DeleteRole"
                ],
                "summary": "response",
                "description": "DeleteRole",
                "operationId": "DeleteRole",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "response",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "success": false,
                                "status": 400,
                                "message": "Bad Request"
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "example": {
                                "status": 200,
                                "message": "Role Deleted Successfully",
                                "data": null
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "addRole": {
            "type": "object",
            "required": [
                "roleName"
            ],
            "properties": {
                "roleName": {
                    "type": "string"
                },
                "permissions": {
                    "type": "Array"
                }
            },
            "example": {
                "roleName":"Customer",
                "description":"Customer",
                "permissions": []
            }
        },
...
    },
...
}